dns_zonefile

Struct DnsZonefile

pub struct DnsZonefile { /* private fields */ }

Implementations

impl DnsZonefile

DnsZonefile This is main struct an interface to generate and parse function which will create Regex resource once using DnsZonefile::default()

use dns_zonefile::{DnsZone, DnsZonefile};
 
fn main() {
    let dns_zonefile = DnsZonefile::default();
    let zonefile_example = r#"
$ORIGIN MYDOMAIN.COM.
$TTL 3600
@    IN    SOA    NS1.NAMESERVER.NET.    HOSTMASTER.MYDOMAIN.COM.    (
            1406291485     ;serial
            3600     ;refresh
            600     ;retry
            604800     ;expire
            86400     ;minimum ttl
)
 
@    NS    NS1.NAMESERVER.NET.
@    NS    NS2.NAMESERVER.NET.
 
    IN  A    2.2.2.2
    A    1.1.1.1
@    A    127.0.0.1
www    A    127.0.0.1
mail    A    127.0.0.1
            A 1.2.3.4
tst 300 IN A 101.228.10.127;this is a comment
 
@    AAAA    ::1
mail    AAAA    2001:db8::1
A  200 AAAA  2001:db8::1
"#;
 
    let Ok(zone_data) = dns_zonefile.parse(zonefile_example) else { return;};
    println!("{}", serde_json::to_string_pretty(&zone_data).unwrap());
 
    let zonefile = dns_zonefile.generate(&zone_data, None);
    println!("{zonefile}");
}

pub fn generate(&self, dns_zone: &DnsZone, template: Option<&str>) -> String

generate Zonefile string from DnsZone data struct like so

pub fn parse(&self, data: &str) -> Result<DnsZone, String>

parse data from zonfile to DnsZone struct

then can convert into json later using serde

THIS function return error for empty zone data rather then all none!